Move internal package to internal dir
