+++ internal/awsiotprotocol/loader_test.go
