+++ internal/awsiotprotocol/mock.go
